requirements.txt
openpyxl==3.0.6
pandas==1.2.0
pysmb==1.2.6
pywin32==300
pywinauto==0.6.8
PyYAML==5.3.1
selenium==3.141.0
urllib3==1.26.2
webdriver_manager==3.5.0